+++ src/main/java/dinkplugin/DinkPlugin.java
@@ -12,13 +12,14 @@ import dinkplugin.notifiers.GroupStorageNotifier;
 import dinkplugin.notifiers.KillCountNotifier;
 import dinkplugin.notifiers.LeaguesNotifier;
 import dinkplugin.notifiers.LevelNotifier;
-import dinkplugin.notifiers.MetaNotifier;
 import dinkplugin.notifiers.LootNotifier;
+import dinkplugin.notifiers.MetaNotifier;
 import dinkplugin.notifiers.PetNotifier;
 import dinkplugin.notifiers.PlayerKillNotifier;
 import dinkplugin.notifiers.QuestNotifier;
 import dinkplugin.notifiers.SlayerNotifier;
 import dinkplugin.notifiers.SpeedrunNotifier;
+import dinkplugin.notifiers.TradeNotifier;
 import dinkplugin.util.Utils;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.ChatMessageType;
@@ -87,6 +88,7 @@ public class DinkPlugin extends Plugin {
     private @Inject GrandExchangeNotifier grandExchangeNotifier;
     private @Inject LeaguesNotifier leaguesNotifier;
     private @Inject MetaNotifier metaNotifier;
+    private @Inject TradeNotifier tradeNotifier;
 
     private final AtomicReference<GameState> gameState = new AtomicReference<>();
 
@@ -114,6 +116,7 @@ public class DinkPlugin extends Plugin {
         killCountNotifier.reset();
         groupStorageNotifier.reset();
         speedrunNotifier.reset();
+        tradeNotifier.reset();
     }
 
     @Provides
@@ -222,6 +225,10 @@ public class DinkPlugin extends Plugin {
                 gambleNotifier.onMesBoxNotification(chatMessage);
                 break;
 
+            case TRADE:
+                tradeNotifier.onTradeMessage(chatMessage);
+                break;
+
             default:
                 // do nothing
                 break;
@@ -282,11 +289,13 @@ public class DinkPlugin extends Plugin {
         speedrunNotifier.onWidgetLoaded(event);
         groupStorageNotifier.onWidgetLoad(event);
         killCountNotifier.onWidget(event);
+        tradeNotifier.onWidgetLoad(event);
     }
 
     @Subscribe
     public void onWidgetClosed(WidgetClosed event) {
         groupStorageNotifier.onWidgetClose(event);
+        tradeNotifier.onWidgetClose(event);
     }
 
     public void addChatSuccess(String message) {